Correction sur l'export Noethsyweb Desk

+++ noethysweb/outils/utils/utils_export_desk.py
@@ -12,6 +12,39 @@ from dbbackup.storage import get_storage
 from core.utils import utils_fichiers
 
 
+class Zip:
+    def __init__(self, nom_zip="", rep_base=""):
+        self.zip_file = zipfile.ZipFile(nom_zip, "w")
+        self.rep_base = rep_base
+
+    def Ajouter(self, nom_fichier):
+        chemin_fichier = os.path.join(self.rep_base, nom_fichier)
+        if os.path.isfile(chemin_fichier):
+            self.Ecrire(chemin_fichier)
+        else:
+            self.Ajouter_repertoire(chemin_fichier)
+
+    def Fermer(self):
+        self.zip_file.close()
+
+    def Ecrire(self, nom_fichier):
+        self.zip_file.write(nom_fichier, nom_fichier.replace(self.rep_base, ""))
+
+    def Ajouter_repertoire(self, folder):
+        for file in os.listdir(folder):
+            full_path = os.path.join(folder, file)
+            if os.path.isfile(full_path):
+                self.Ecrire(full_path)
+            elif os.path.isdir(full_path):
+                self.Ajouter_repertoire(full_path)
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.Fermer()
+
+
 class Desk:
     nom_fichier = "noethyswebdesk.7z"
     repertoire = "desk"
@@ -38,37 +71,61 @@ class Desk:
 
     def Generer(self, mdp=None):
         """ Génère un fichier desk au format 7z """
+        logger.debug("Génération du fichier desk...")
+
         # Vide le répertoire temp
+        logger.debug("Vidage des fichiers temporaires...")
         call_command("reset_rep_temp")
+        self.Nettoyer_fichiers_temporaires()
 
         # Création du zip data
+        logger.debug("Génération du zip data...")
         call_command("dumpdatautf8", format="json", indent=4, output="data.json", exclude=["core.Historique", "contenttypes"])
         with zipfile.ZipFile("data.zip", "w", zipfile.ZIP_DEFLATED) as myzip:
             myzip.write("data.json")
+        logger.debug("Génération du zip data terminé.")
 
         # Création du zip media
-        shutil.make_archive(settings.MEDIA_ROOT, format="zip", root_dir=settings.MEDIA_ROOT)
+        logger.debug("Génération du zip media...")
+        liste_repertoires = os.listdir(settings.MEDIA_ROOT)
+        liste_exclusions = ("django-summernote", "photos", "pieces_jointes", "temp")
+        with Zip("media.zip", settings.MEDIA_ROOT) as zip:
+            for nom_rep in liste_repertoires:
+                if nom_rep not in liste_exclusions:
+                    zip.Ajouter(nom_rep)
+        logger.debug("Génération du zip media terminé.")
 
         # Création du fichier des préférences
+        logger.debug("Génération du fichier des préférences...")
         with open("preferences.json", "w") as fichier:
             data = {"URL_GESTION": settings.URL_GESTION}
             json.dump(data, fichier)
+        logger.debug("Génération du fichier des préférences terminé.")
 
         # Création du 7z global
+        logger.debug("Génération du 7zip...")
         import py7zr
         with py7zr.SevenZipFile(self.Get_chemin_fichier_7z(), "w", password=mdp or settings.SECRET_EXPORT_DESK) as archive:
             archive.write("data.zip")
             archive.write("media.zip")
             archive.write("preferences.json")
-            archive.write("core/desk/run.py", arcname="run.py")
-            archive.write("core/desk/lisezmoi.txt", arcname="lisezmoi.txt")
-            archive.write("core/desk/lib/settings_modele.py", arcname="lib/settings_modele.py")
-            archive.write("core/desk/lib/requirements.txt", arcname="lib/requirements.txt")
+            archive.write(os.path.join(settings.BASE_DIR, "core/desk/run.py"), arcname="run.py")
+            archive.write(os.path.join(settings.BASE_DIR, "core/desk/lisezmoi.txt"), arcname="lisezmoi.txt")
+            archive.write(os.path.join(settings.BASE_DIR, "core/desk/lib/settings_modele.py"), arcname="lib/settings_modele.py")
+            archive.write(os.path.join(settings.BASE_DIR, "core/desk/lib/requirements.txt"), arcname="lib/requirements.txt")
+        logger.debug("Génération du 7zip terminé.")
 
         # Nettoyage
         time.sleep(2)
+        logger.debug("Nettoyage des fichiers temporaires...")
+        self.Nettoyer_fichiers_temporaires()
+
+        logger.debug("Génération du fichier desk terminé.")
+
+    def Nettoyer_fichiers_temporaires(self):
         for nom_fichier in ("data.json", "data.zip", "media.zip", "preferences.json"):
-            os.remove(nom_fichier)
+            if os.path.isfile(nom_fichier):
+                os.remove(nom_fichier)
 
     def Get_storage(self):
         """ Ouverture du storage """
@@ -81,11 +138,14 @@ class Desk:
         storage = self.Get_storage()
 
         # Lecture du contenu du fichier
+        logger.debug("Lecture du fichier à envoyer vers le storage...")
         with open(self.Get_chemin_fichier_7z(), "rb") as f:
             contenu_fichier = f.read()
 
         # Envoi du fichier vers le storage
+        logger.debug("Envoi du fichier vers le storage...")
         storage.storage.save(name=self.nom_fichier, content=ContentFile(contenu_fichier, name=self.nom_fichier))
+        logger.debug("Envoi du fichier vers le storage terminé.")
 
     def Nettoyer_storage(self):
         """ Supprime le fichier desk sur le storage """
